Fix all greenplum_schedule tests diffs in single node mode
Since cases in greenplum_schedule are designed to test features
provided in greenplum, and many of those features are natually not
supported in single node mode. So we only select a few of those tests
to run.

And in this commit, we select 75 cases of the  greenplum_schedule and
make them pass.